+++ src/settingsmanager.h
@@ -12,6 +12,14 @@ class SettingsManager : public QObject {
       QString hostname READ hostname WRITE setHostname NOTIFY hostnameChanged)
   Q_PROPERTY(bool autostartEnabled READ autostartEnabled NOTIFY
                  autostartEnabledChanged)
+  Q_PROPERTY(int windowWidth READ windowWidth WRITE setWindowWidth NOTIFY
+                 windowWidthChanged)
+  Q_PROPERTY(int windowHeight READ windowHeight WRITE setWindowHeight NOTIFY
+                 windowHeightChanged)
+  Q_PROPERTY(int windowX READ windowX WRITE setWindowX NOTIFY windowXChanged)
+  Q_PROPERTY(int windowY READ windowY WRITE setWindowY NOTIFY windowYChanged)
+  Q_PROPERTY(QString buildHost READ buildHost WRITE setBuildHost NOTIFY
+                 buildHostChanged)
 
 public:
   explicit SettingsManager(QObject *parent = nullptr);
@@ -27,10 +35,30 @@ public:
   Q_INVOKABLE void setAutostartEnabled(bool enabled);
   Q_INVOKABLE void checkAndCreateAutostart();
 
+  int windowWidth() const { return m_windowWidth; }
+  void setWindowWidth(int width);
+
+  int windowHeight() const { return m_windowHeight; }
+  void setWindowHeight(int height);
+
+  int windowX() const { return m_windowX; }
+  void setWindowX(int x);
+
+  int windowY() const { return m_windowY; }
+  void setWindowY(int y);
+
+  QString buildHost() const { return m_buildHost; }
+  void setBuildHost(const QString &buildHost);
+
 signals:
   void startHiddenChanged();
   void hostnameChanged();
   void autostartEnabledChanged();
+  void windowWidthChanged();
+  void windowHeightChanged();
+  void windowXChanged();
+  void windowYChanged();
+  void buildHostChanged();
 
 private:
   void loadSettings();
@@ -44,6 +72,11 @@ private:
 
   bool m_startHidden;
   QString m_hostname;
+  int m_windowWidth;
+  int m_windowHeight;
+  int m_windowX;
+  int m_windowY;
+  QString m_buildHost;
 };
 
 #endif // SETTINGSMANAGER_H